+++ xdsl/dialects/eqsat.py
@@ -13,7 +13,7 @@ from collections.abc import Sequence
 from typing import ClassVar
 
 from xdsl.dialects.builtin import IntAttr
-from xdsl.ir import Attribute, Block, Dialect, Region, SSAValue
+from xdsl.ir import Attribute, Dialect, Region, SSAValue
 from xdsl.irdl import (
     AnyAttr,
     IRDLOperation,
@@ -23,11 +23,9 @@ from xdsl.irdl import (
     opt_attr_def,
     region_def,
     result_def,
-    successor_def,
     traits_def,
     var_operand_def,
     var_result_def,
-    var_successor_def,
 )
 from xdsl.traits import HasParent, IsTerminator, Pure, SingleBlockImplicitTerminator
 from xdsl.utils.exceptions import DiagnosticException, VerifyException
@@ -128,36 +126,11 @@ class YieldOp(IRDLOperation):
         super().__init__(operands=[values])
 
 
-@irdl_op_definition
-class ChooseOp(IRDLOperation):
-    """
-    This operation can be used in pdl_interp matchers and
-    integrates with the backtracking mechanism. It holds multiple
-    "choices" (successors). When this operation is encountered,
-    a BacktrackPoint is stored, and the choice is visited.
-    When this execution of this choice eventually finalizes, the
-    backtracking logic will jump to the next choice, until all
-    choices are exhausted. Finally, the default successor is visited.
-    """
-
-    name = "eqsat.choose"
-    default_dest = successor_def()
-    choices = var_successor_def()
-    traits = traits_def(IsTerminator())
-    assembly_format = "`from` $choices `then` $default_dest attr-dict"
-
-    def __init__(self, choices: Sequence[Block], default: Block):
-        super().__init__(
-            successors=[default, choices],
-        )
-
-
 EqSat = Dialect(
     "eqsat",
     [
         EClassOp,
         YieldOp,
         EGraphOp,
-        ChooseOp,
     ],
 )
